refactor: station의 패키지 위치 수정
domain-model로 이동
