+++ src/main/java/subway/Application.java
@@ -1,10 +1,16 @@
 package subway;
 
+import subway.manager.MainManager;
+import subway.manager.Manager;
+import subway.util.DummyData;
+
 import java.util.Scanner;
 
 public class Application {
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
-        // TODO: 프로그램 구현
+        DummyData.init();
+        Manager manager = new MainManager();
+        manager.execute(scanner);
     }
 }
