+++ src/test/java/subway/domain/section/model/SectionTest.java
@@ -2,7 +2,7 @@ package subway.domain.section.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import subway.station.model.Station;
+import subway.domain.station.model.Station;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
