+++ src/test/java/subway/domain/section/service/SectionServiceTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import subway.domain.section.model.Section;
 import subway.domain.section.model.SectionRepository;
-import subway.station.model.Station;
+import subway.domain.station.model.Station;
 
 import java.util.List;
